fixed the import issue again
import cycle error fixed with the issue requirements by moving Airdrop() to the blockchain package
