+++ internal/genesis/genesis.go
@@ -5,14 +5,12 @@ import (
 	"crypto/ecdsa"
 	"crypto/elliptic"
 	"crypto/rand"
-	"database/sql"
 	"encoding/hex"
 	"errors"
 	"io"
 
 	"os"
 
-	"github.com/SIGBlockchain/project_aurum/internal/blockchain"
 	"github.com/SIGBlockchain/project_aurum/internal/constants"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/block"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/contracts"
@@ -58,75 +56,6 @@ func BringOnTheGenesis(genesisPublicKeyHashes [][]byte, initialAurumSupply uint6
 
 	return genesisBlock, nil
 }
-
-func Airdrop(blockchainz string, metadata string, accountBalanceTable string, genesisBlock block.Block) error {
-	// create blockchain file
-	file, err := os.Create(blockchainz)
-	if err != nil {
-		return errors.New("Failed to create blockchain file")
-	}
-	file.Close()
-
-	// create metadata file
-	file, err = os.Create(metadata)
-	if err != nil {
-		return errors.New("Failed to create metadata table")
-	}
-	file.Close()
-
-	// open metadata file and create the table
-	db, err := sql.Open("sqlite3", metadata)
-	if err != nil {
-		return errors.New("Failed to open table")
-	}
-
-	_, err = db.Exec("CREATE table METADATA (height INTEGER PRIMARY KEY, position INTEGER, size INTEGER, hash TEXT)")
-	if err != nil {
-		return errors.New("Failed to create table")
-	}
-	db.Close()
-
-	// add genesis block into blockchain
-	err = blockchain.AddBlock(genesisBlock, blockchainz, metadata)
-	if err != nil {
-		return errors.New("Failed to add genesis block into blockchain")
-	}
-
-	// create accounts file
-	file, err = os.Create(accountBalanceTable)
-	if err != nil {
-		return errors.New("Failed to create accounts table")
-	}
-	file.Close()
-
-	accDb, err := sql.Open("sqlite3", accountBalanceTable)
-	if err != nil {
-		return errors.New("Failed to open newly created accounts db")
-	}
-	defer accDb.Close()
-
-	_, err = accDb.Exec("CREATE TABLE account_balances (public_key_hash TEXT, balance INTEGER, nonce INTEGER)")
-	if err != nil {
-		return errors.New("Failed to create acount_balances table")
-	}
-
-	stmt, err := accDb.Prepare("INSERT INTO account_balances VALUES (?, ?, ?)")
-	if err != nil {
-		return errors.New("Failed to create statement for inserting into account table")
-	}
-
-	for _, contrcts := range genesisBlock.Data {
-		var contract contracts.Contract
-		contract.Deserialize(contrcts)
-		_, err := stmt.Exec(hex.EncodeToString(contract.RecipPubKeyHash), contract.Value, 0)
-		if err != nil {
-			return errors.New("Failed to execute statement for inserting into account table")
-		}
-	}
-	return nil
-}
-
-// var genesisHashFile = "genesis_hashes.txt"
 
 // Open the genesisHashFile
 // Read line by line
