+++ internal/genesis/genesis_test.go
@@ -1,17 +1,16 @@
 package genesis
 
 import (
-	"bytes"
 	"crypto/ecdsa"
 	"crypto/elliptic"
 	"crypto/rand"
 	"database/sql"
 	"encoding/hex"
-	"io/ioutil"
 	"os"
 	"reflect"
 	"testing"
 
+	"github.com/SIGBlockchain/project_aurum/internal/blockchain"
 	"github.com/SIGBlockchain/project_aurum/internal/constants"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/block"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/contracts"
@@ -84,87 +83,6 @@ func TestBringOnTheGenesis(t *testing.T) {
 	}
 }
 
-func TestAirdrop(t *testing.T) {
-	var pkhashes [][]byte
-	for i := 0; i < 100; i++ {
-		someKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-		someKeyPKHash := hashing.New(publickey.Encode(&someKey.PublicKey))
-		pkhashes = append(pkhashes, someKeyPKHash)
-	}
-	genny, _ := BringOnTheGenesis(pkhashes, 1000)
-	type args struct {
-		blockchain   string
-		metadata     string
-		accounts     string
-		genesisBlock block.Block
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{
-			args: args{
-				blockchain:   "blockchain.dat",
-				metadata:     constants.MetadataTable,
-				accounts:     constants.AccountsTable,
-				genesisBlock: genny,
-			},
-		},
-	}
-	for _, tt := range tests {
-		defer func() {
-			os.Remove(tt.args.metadata)
-			os.Remove(tt.args.blockchain)
-			os.Remove(tt.args.accounts)
-		}()
-		t.Run(tt.name, func(t *testing.T) {
-			if err := Airdrop(tt.args.blockchain, tt.args.metadata, tt.args.accounts, tt.args.genesisBlock); (err != nil) != tt.wantErr {
-				t.Errorf("Airdrop() error = %v, wantErr %v", err, tt.wantErr)
-			}
-			fileGenny, err := ioutil.ReadFile(tt.args.blockchain)
-			if err != nil {
-				t.Errorf("Failed to open file" + err.Error())
-			}
-			serializedGenny := genny.Serialize()
-			if !bytes.Equal(fileGenny[4:], serializedGenny) {
-				t.Errorf("Genesis block does not match file block")
-			}
-
-			db, err := sql.Open("sqlite3", tt.args.accounts)
-			if err != nil {
-				t.Errorf("Failed to open accounts table: " + err.Error())
-			}
-			defer db.Close()
-
-			rows, err := db.Query(sqlstatements.GET_PUB_KEY_HASH_BALANCE_NONCE_FROM_ACCOUNT_BALANCES)
-			if err != nil {
-				t.Errorf("failed to create rows for queries")
-			}
-			defer rows.Close()
-
-			var pkhCount int
-			var pkhStr string
-			var balance int
-			var nonce int
-			for rows.Next() {
-				rows.Scan(&pkhStr, &balance, &nonce)
-				pkhash, _ := hex.DecodeString(pkhStr)
-				if !bytes.Equal(pkhash, pkhashes[pkhCount]) {
-					t.Errorf("hashes don't match: %v != %v\n", pkhash, pkhashes[pkhCount])
-				}
-				if balance != 10 {
-					t.Errorf("balance does not match: %v != %v\n", balance, 10)
-				}
-				if nonce != 0 {
-					t.Errorf("nonce does not match: %v != %v\n", nonce, 0)
-				}
-				pkhCount++
-			}
-		})
-	}
-}
-
 func TestReadGenesisHashes(t *testing.T) {
 	GenerateGenesisHashFile(50)
 
@@ -224,7 +142,7 @@ func TestGenesisReadsAppropriately(t *testing.T) {
 	}
 
 	//airdrop
-	err = Airdrop("blockchain.dat", constants.MetadataTable, constants.AccountsTable, genesisBlock)
+	err = blockchain.Airdrop("blockchain.dat", constants.MetadataTable, constants.AccountsTable, genesisBlock)
 	if err != nil {
 		t.Errorf("failed to airdrop genesis block: %s", err.Error())
 	}
